Configurable masking and hiding for the env endpoint
- Changes the env endpoint to be disabled by default (requires user to enable in config)
- If enabled, all env endpoint values are masked with *****
- Allow user to add a bean implementing EnvironmentEndpointFilter which:
  - allows legacy operation to be allowed so it behaves as before
  - can be set to mask or reveal all
  - allows patterns for masking to be defined if reveal all is set
  - allows patterns for plain-text values if mask all is set

The call to the filter also contains the Principal if there is one, so the user can choose
to alter visibility of items based on the currently logged in user
