+++ src/main/docs/guide/appendix/breaks.adoc
@@ -1,5 +1,30 @@
 This section documents breaking changes between Micronaut versions
 
+== 3.3.0
+
+- The link:#environmentEndpoint[environmental endpoint] is now disabled by default. To enable it, you must update your endpoint config:
+
+[source,yaml]
+----
+endpoints:
+  env:
+    enabled: true
+----
+
+This will then be available, but mask all values.  To restore the legacy functionality, you can add a Bean like so:
+
+.Legacy Environment Filtering Bean
+[source,java]
+----
+@Singleton
+public class LegacyEnvEndpointFilter implements EnvironmentEndpointFilter {
+    @Override
+    public void specifyFiltering(EnvironmentFilterSpecification specification) {
+        specification.legacyFiltering();
+    }
+}
+----
+
 == 3.2.0
 
 - The HTTP client now does SSL certificate verification by default. The old insecure behavior can be re-enabled by setting the `micronaut.http.client.ssl.insecureTrustAllCertificates` property to `true`, but consider using a trust store instead if you're using self-signed certificates.
